style(dataset): correct selenium folders and file names
#150
