+++ app/modules/dataset/tests/test_selenium_ide/Test_UI.py
@@ -1,0 +1,25 @@
+# Generated by Selenium IDE
+from selenium.webdriver.common.by import By
+from core.selenium.common import initialize_driver, close_driver
+
+
+class TestUI:
+    def setup_method(self, method):
+        self.driver = initialize_driver()
+        self.vars = {}
+
+    def teardown_method(self, method):
+        close_driver(self.driver)
+
+    def test_selenium(self):
+        self.driver.get("http://localhost:5000/")
+        self.driver.set_window_size(1536, 816)
+        self.driver.find_element(By.LINK_TEXT, "Cats").click()
+        self.driver.find_element(By.ID, "btnGroupDropExport3").click()
+        self.driver.find_element(By.ID, "btnGroupDropExport3").click()
+        self.driver.find_element(By.ID, "btnGroupDrop1").click()
+        self.driver.find_element(By.LINK_TEXT, "Syntax check").click()
+        self.driver.find_element(
+            By.CSS_SELECTOR, ".list-group-item:nth-child(2) .col-12 > .btn > .feather"
+        ).click()
+        self.driver.close()
